refactor: 共通関数をutilへ
