+++ src/slash-command/nita-list.js
@@ -4,6 +4,7 @@ import { planetScaleRepository } from '../infra/repository/planetscale.js';
 import { InteractionResponseType, MessageFlags } from 'discord-api-types/v10';
 import { validateSlashCommand } from '../util/validate-slash-command.js';
 import { getByCode } from '../const/track.js';
+import { displayMilliseconds } from '../util/time.js';
 
 /** @type { import('../types').SlashCommand } */
 export default {
@@ -70,15 +71,4 @@ export default {
       };
     }
   },
-};
-
-/**
- * @param {number} milliseconds
- */
-const displayMilliseconds = (milliseconds) => {
-  const millisecondsStr = Math.floor(milliseconds % 1000).toString().padStart(3, '0');
-  const seconds = Math.floor(milliseconds / 1000);
-  const secondsStr = (seconds % 60).toString().padStart(2, '0');
-  const minutes = Math.floor(seconds / 60);
-  return `${minutes}:${secondsStr}.${millisecondsStr}`;
 };
